Move Passkey related methods to Passkeys service
